requirements.txt
Django
django-debug-toolbar
django-tinymce
gunicorn
Pillow
